# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Leduc Receipt Pro API",
    "version": "2.0.0",
    "description": "Internal API routes for receipt scanning, CRA exports, Stripe payments, QBO/Xero integration, team management, email ingestion, and health checks. Consumed by the Next.js client and external webhooks — not intended for public use."
  },
  "servers": [{ "url": "/api", "description": "Same-origin API" }],
  "paths": {
    "/api/cra/generate": {
      "get": {
        "summary": "Generate CRA Audit PDF Package",
        "description": "Generates a multi-page PDF with T2125/T777 pre-fill data, expense breakdowns, mileage, and top vendors. Accepts tax year or date range to scope data.",
        "parameters": [
          { "in": "query", "name": "year", "schema": { "type": "integer", "minimum": 2000, "maximum": 2099 }, "description": "Tax year (defaults to last year)" },
          { "in": "query", "name": "fromDate", "schema": { "type": "string", "format": "date" }, "description": "Start date (YYYY-MM-DD) to override year-based range" },
          { "in": "query", "name": "toDate", "schema": { "type": "string", "format": "date" }, "description": "End date (YYYY-MM-DD) to override year-based range" }
        ],
        "responses": {
          "200": { "description": "PDF binary", "content": { "application/pdf": {} } },
          "400": { "description": "Invalid parameters" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Organization not found" },
          "500": { "description": "Generation failed" }
        }
      }
    },
...
    "/api/digest/missing-receipts": {
      "get": {
        "summary": "Daily digest of missing receipts",
        "description": "CRON-triggered endpoint that generates an email digest listing bank transactions without matching receipts for the previous day. Authorized via CRON_SECRET.",
        "parameters": [
          { "in": "query", "name": "token", "schema": { "type": "string" }, "required": true, "description": "CRON_SECRET for authorization" }
        ],
        "responses": {
          "200": { "description": "Digest sent" },
          "401": { "description": "Invalid CRON_SECRET" },
          "500": { "description": "Digest generation failed" }
        }
      }
    },
...
  }
}